requirements.txt
alembic==1.4.2
appdirs==1.4.3
attrs==19.3.0
awscli==1.18.54
black==19.10b0
botocore==1.16.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
docutils==0.15.2
face-detection==0.1.4
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
gdown==3.11.0
gunicorn==20.0.4
idna==2.9
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
kiwisolver==1.2.0
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.0.3
pathspec==0.8.0
Pillow==7.1.2
psycopg2-binary==2.8.5
pusher==3.0.0
pyasn1==0.4.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.0
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
easydict
tensorflow==2.1.0